requirements.txt
certifi==2016.8.8
future==0.15.2
python-telegram-bot==5.0.0
redis==2.10.5
urllib3==1.16
